refactored storages, node storages, tests and so on because there's never enough refactoring in one's life
