+++ tests/proofs/conftest.py
@@ -3,16 +3,13 @@ import logging
 import pytest
 
 from aitools.proofs.knowledge_base import KnowledgeBase
-from aitools.storage.dummy import DummyLogicObjectStorage, DummyIndexedLogicObjectStorage
-from aitools.storage.inmem_serializing import InMemSerializingLogicObjectStorage, \
-    DummyIndexedSerializingLogicObjectStorage
+from tests.implementations import storage_implementations
 
 logging.basicConfig(format="[%(levelname)s] %(name)s - %(message)s")
 logging.getLogger().setLevel(logging.WARNING)
 logging.getLogger('txn').setLevel(logging.WARNING)
 
 
-@pytest.fixture(params=[DummyLogicObjectStorage, DummyIndexedLogicObjectStorage, InMemSerializingLogicObjectStorage,
-                        DummyIndexedSerializingLogicObjectStorage])
+@pytest.fixture(params=storage_implementations)
 def TestKnowledgeBase(request):
     return lambda: KnowledgeBase(request.param())
