+++ tests/storage/test_storage.py
@@ -6,30 +6,16 @@ import pytest
 from aitools.logic import Substitution
 from aitools.logic.utils import constants, VariableSource, normalize_variables
 from aitools.storage.base import LogicObjectStorage
-from aitools.storage.dummy import DummyLogicObjectStorage, DummyIndexedLogicObjectStorage
-from aitools.storage.inmem_serializing import InMemSerializingLogicObjectStorage, \
-    DummyIndexedSerializingLogicObjectStorage
+from tests import implementations
 
 
-@dataclass
-class StorageImplementation:
-    storage_factory: Callable[[], LogicObjectStorage]
-    preserves_identity: bool
-
-@pytest.fixture(
-    params=[
-        StorageImplementation(DummyLogicObjectStorage, True),
-        StorageImplementation(InMemSerializingLogicObjectStorage, False),
-        StorageImplementation(DummyIndexedLogicObjectStorage, True),
-        StorageImplementation(DummyIndexedSerializingLogicObjectStorage, False)
-    ])
+@pytest.fixture(params=implementations.storage_implementations)
 def storage_implementation(request):
-    """Each param is a pair of a storage factory and a boolean which says if it preserves identity"""
     return request.param
 
 
-def test_retrieve_known_formula(storage_implementation: StorageImplementation):
-    storage: LogicObjectStorage = storage_implementation.storage_factory()
+def test_retrieve_known_formula(storage_implementation):
+    storage: LogicObjectStorage = storage_implementation()
 
     IsA, dylan, cat = constants('IsA, dylan, cat')
 
@@ -45,14 +31,9 @@ def test_retrieve_known_formula(storage_implementation: StorageImplementation):
     assert isinstance(unifier, Substitution)
     assert retrieved == IsA(dylan, cat)
 
-    if storage_implementation.preserves_identity:
-        assert retrieved is formula
-    else:
-        assert retrieved is not formula
 
-
-def test_retrieve_known_open_formula(storage_implementation: StorageImplementation):
-    storage: LogicObjectStorage = storage_implementation.storage_factory()
+def test_retrieve_known_open_formula(storage_implementation):
+    storage: LogicObjectStorage = storage_implementation()
 
     v = VariableSource()
 
@@ -67,8 +48,8 @@ def test_retrieve_known_open_formula(storage_implementation: StorageImplementati
     assert retrieved == {IsA(dylan, cat), IsA(hugo, cat)}
 
 
-def test_normalized_formulas_added_only_once(storage_implementation: StorageImplementation):
-    storage: LogicObjectStorage = storage_implementation.storage_factory()
+def test_normalized_formulas_added_only_once(storage_implementation):
+    storage: LogicObjectStorage = storage_implementation()
 
     v = VariableSource()
     Foo, a, b = constants('Foo, a, b')
